Feature added: Take picture of the place and save it.
Added the image package by Flutter and used ImagePicker to ask permission for camera access and let the user take an image.

When taken the image is displayed as preview in the form with the option to tap it and retake the image.

When submitted, an avatar container is next to the place title which displays the image's content.

In place details, the image takes the full screen.