Added a new feature: Getting current location.
Using the Flutter location package, created a button that would get the user's current location after asking for permission. Another button to open Google Maps and choose a location manually. Above them a container that will display a preview of the chosen location.

Limitations: I couldn't use third party packages related to Google Maps due to the need of a Google Maps Account, so the Button to choose a location manuaaly on the map and the preview aren't implemented.