Last featue: Storing the data in the local storage
Using 3 third-party packages: path_provider to get the system's directory that is for the app allocated by the OS. path to manipulate paths easily. sqflite for creating a database in the folder allocated by the OS.

In the provider, added the functionality to store the image, title, and the location data into locale storage (most of the code is by the documentation of the packages). In addition created a function outside the provider's class to create the database user_places where all the user's data is stored.

In the main screen (home screen), implemented the FutureBuilder to keep track of future variables changes which is a consumer that listens to the loadData function.